relocate badge components under gamification directory
* standardize badge names
* remove unused test components
