--- timApp/static/scripts/tim/gamification/badge-testi-component.ts
@@ -1,1 +1,0 @@
-// TODO
