+++ timApp/static/scripts/tim/gamification/badge/badge-creator.component.scss
