+++ timApp/static/scripts/tim/gamification/badge/badge-giver.component.scss
