+++ timApp/static/scripts/tim/gamification/badge/badge-viewer.component.scss
