+++ timApp/static/scripts/tim/gamification/badge/badge-viewer.component.ts
@@ -2,12 +2,11 @@ import type {OnInit} from "@angular/core";
 import {Component, NgModule} from "@angular/core";
 import {CommonModule} from "@angular/common";
 import {FormsModule} from "@angular/forms";
-import {Users} from "tim/user/userService";
+import {Users} from "../../user/userService";
 import {HttpClient, HttpClientModule} from "@angular/common/http";
-import {BadgeModule} from "tim/Badge/Badge-component";
-import {BadgeTestModule} from "tim/Badge/badge-test-component";
-import {BadgeService} from "tim/Badge/badge.service";
-import type {IBadge} from "tim/Badge/badge.interface";
+import {BadgeModule} from "./badge.component";
+import {BadgeService} from "./badge.service";
+import type {IBadge} from "./badge.interface";
 import {Subscription} from "rxjs";
 
 @Component({
@@ -32,7 +31,7 @@ import {Subscription} from "rxjs";
             </ng-container>
         </div>
         `,
-    styleUrls: ["badge-viewer-component.scss"],
+    styleUrls: ["badge-viewer.component.scss"],
 })
 export class BadgeViewerComponent implements OnInit {
     userName?: string;
@@ -44,7 +43,7 @@ export class BadgeViewerComponent implements OnInit {
     constructor(private http: HttpClient, private badgeService: BadgeService) {}
 
     /**
-     * Tyhjentää badge -taulukon ja kutsuu Badge-servicen metodia joka hakee käyttäjälle kuuluvat badget.
+     * Tyhjentää badge -taulukon ja kutsuu badge-servicen metodia joka hakee käyttäjälle kuuluvat badget.
      * @param id käyttäjän id (tällähetkellä käytetään sisäänkirjautuneen käyttäjän ID:tä)
      */
     async getBadges(id: number) {
@@ -80,13 +79,7 @@ export class BadgeViewerComponent implements OnInit {
 
 @NgModule({
     declarations: [BadgeViewerComponent],
-    imports: [
-        CommonModule,
-        FormsModule,
-        HttpClientModule,
-        BadgeModule,
-        BadgeTestModule,
-    ],
+    imports: [CommonModule, FormsModule, HttpClientModule, BadgeModule],
     exports: [BadgeViewerComponent],
 })
 export class BadgeViewerModule {}
