+++ timApp/static/scripts/tim/gamification/badge/badge.component.scss
