+++ timApp/static/scripts/tim/gamification/badge/badge.interface.ts
